+++ datar/base/funs.py
@@ -3,527 +3,16 @@
 If a function uses DataFrame/DataFrameGroupBy as first argument, it may be
 registered by `register_verb` and should be placed in `./verbs.py`
 """
-import builtins
-import math
-import datetime
 import itertools
-import functools
-from typing import Any, Iterable, List, Optional, Type, Union
+from typing import Any, Iterable, Optional
 
-import numpy
 import pandas
-from pandas.core.dtypes.common import (
-    is_categorical_dtype, is_float_dtype, is_int64_dtype, is_integer_dtype,
-    is_numeric_dtype, is_string_dtype
-)
-from pandas import Series, Categorical, DataFrame
-from pandas.core.groupby.generic import SeriesGroupBy
-from pipda import Context, register_func
+from pandas import Categorical, DataFrame
+from pipda import register_func
 
-from .na import NA
-from ..core.utils import categorized, get_option, logger, Array, length_of
 from ..core.middlewares import WithDataEnv
-from ..core.collections import Collection
-from ..core.types import (
-    BoolOrIter, CategoricalLikeType, DoubleOrIter, IntOrIter,
-    NumericOrIter,
-    NumericType, SeriesLikeType, StringOrIter, is_scalar_int, IntType,
-    is_iterable, is_series_like, is_scalar, is_null
-)
+from ..core.types import NumericType
 from ..core.contexts import Context
-
-# pylint: disable=redefined-builtin,invalid-name
-
-@functools.singledispatch
-def _as_date_dummy(
-        x: Any,
-        format: Optional[str] = None,
-        try_formats: Optional[List[str]] = None,
-        optional: bool = False,
-        tz: Union[IntType, datetime.timedelta] = 0,
-        origin: Any = None
-) -> datetime.date:
-    """Convert a dummy object to date"""
-    raise ValueError(f'Unable to convert to date with type: {type(x)!r}')
-
-@_as_date_dummy.register(datetime.date)
-def _(
-        x: datetime.date,
-        format: Optional[str] = None,
-        try_formats: Optional[List[str]] = None,
-        optional: bool = False,
-        tz: Union[int, datetime.timedelta] = 0,
-        origin: Any = None
-) -> datetime.date:
-    if is_scalar_int(tz):
-        tz = datetime.timedelta(hours=int(tz))
-
-    return x + tz
-
-@_as_date_dummy.register(datetime.datetime)
-def _(
-        x: datetime.datetime,
-        format: Optional[str] = None,
-        try_formats: Optional[List[str]] = None,
-        optional: bool = False,
-        tz: Union[IntType, datetime.timedelta] = 0,
-        origin: Any = None
-):
-    if is_scalar_int(tz):
-        tz = datetime.timedelta(hours=int(tz))
-
-    return (x + tz).date()
-
-@_as_date_dummy.register(str)
-def _(
-        x: str,
-        format: Optional[str] = None,
-        try_formats: Optional[List[str]] = None,
-        optional: bool = False,
-        tz: Union[IntType, datetime.timedelta] = 0,
-        origin: Any = None
-) -> datetime.date:
-    if is_scalar_int(tz):
-        tz = datetime.timedelta(hours=int(tz))
-
-    try_formats = try_formats or [
-        "%Y-%m-%d",
-        "%Y/%m/%d",
-        "%Y-%m-%d %H:%M:%S",
-        "%Y/%m/%d %H:%M:%S"
-    ]
-    if not format:
-        format = try_formats
-    else:
-        format = [format]
-
-    for fmt in format:
-        try:
-            return (datetime.datetime.strptime(x, fmt) + tz).date()
-        except ValueError:
-            continue
-
-    if optional:
-        return numpy.nan
-
-    raise ValueError(
-        "character string is not in a standard unambiguous format"
-    )
-
-
-@_as_date_dummy.register(Series)
-def _(
-        x: Series,
-        format: Optional[str] = None,
-        try_formats: Optional[List[str]] = None,
-        optional: bool = False,
-        tz: Union[IntType, datetime.timedelta] = 0,
-        origin: Any = None
-) -> datetime.date:
-    return _as_date_dummy(
-        x.values[0],
-        format=format,
-        try_formats=try_formats,
-        optional=optional,
-        tz=tz,
-        origin=origin
-    )
-
-@_as_date_dummy.register(int)
-@_as_date_dummy.register(numpy.integer)
-def _(
-        x: IntType,
-        format: Optional[str] = None,
-        try_formats: Optional[List[str]] = None,
-        optional: bool = False,
-        tz: Union[IntType, datetime.timedelta] = 0,
-        origin: Any = None
-) -> datetime.date:
-    if isinstance(tz, (int, numpy.integer)):
-        tz = datetime.timedelta(hours=int(tz))
-
-    if isinstance(origin, str):
-        origin = _as_date_dummy(origin)
-
-    dt = origin + datetime.timedelta(days=int(x)) + tz
-
-    if isinstance(dt, datetime.datetime):
-        return dt.date()
-    return dt
-
-@register_func(None, context=Context.EVAL)
-def as_date(
-        x: DataFrame,
-        format: Optional[str] = None,
-        try_formats: Optional[List[str]] = None,
-        optional: bool = False,
-        tz: Union[IntType, datetime.timedelta] = 0,
-        origin: Any = None
-) -> Iterable[datetime.date]:
-    """Convert an object to a datetime.date object
-
-    See: https://rdrr.io/r/base/as.Date.html
-
-    Args:
-        x: Object that can be converted into a datetime.date object
-        format:  If not specified, it will try try_formats one by one on
-            the first non-NA element, and give an error if none works.
-            Otherwise, the processing is via strptime
-        try_formats: vector of format strings to try if format is not specified.
-        optional: indicating to return NA (instead of signalling an error)
-            if the format guessing does not succeed.
-        origin: a datetime.date/datetime object, or something which can be
-            coerced by as_date(origin, ...) to such an object.
-        tz: a time zone offset or a datetime.timedelta object.
-            Note that time zone name is not supported yet.
-
-    Returns:
-        The datetime.date object
-    """
-    if not isinstance(x, (Series, SeriesGroupBy)):
-        x = Series([x]) if is_scalar(x) else Series(x)
-
-    return x.transform(
-        _as_date_dummy,
-        format=format,
-        try_formats=try_formats,
-        optional=optional,
-        tz=tz,
-        origin=origin
-    )
-
-def _as_type(x: Any, type_: Type) -> Any:
-    """Convert x or elements of x to certain type"""
-    if is_series_like(x):
-        return x.astype(type_)
-
-    if is_scalar(x):
-        return type_(x)
-
-    return type(x)(map(type_, x))
-
-@register_func(None, context=Context.EVAL)
-def as_character(x: Any) -> StringOrIter:
-    """Convert an object or elements of an iterable into string
-
-    Args:
-        x: The object
-
-    Returns:
-        When x is an array or a series, return x.astype(str).
-        When x is iterable, convert elements of it into strings
-        Otherwise, convert x to string.
-    """
-    return _as_type(x, str)
-
-@register_func(None, context=Context.EVAL)
-def as_double(x: Any) -> DoubleOrIter:
-    """Convert an object or elements of an iterable into double/float
-
-    Args:
-        x: The object
-
-    Returns:
-        When x is an array or a series, return x.astype(float).
-        When x is iterable, convert elements of it into floats
-        Otherwise, convert x to float.
-    """
-    return _as_type(x, numpy.double)
-
-@register_func(None, context=Context.EVAL)
-def as_factor(x: Iterable[Any]) -> Categorical:
-    """Convert an iterable into a pandas.Categorical object
-
-    Args:
-        x: The iterable
-
-    Returns:
-        The converted categorical object
-    """
-    return Categorical(x)
-
-as_categorical = as_factor
-
-@register_func(None, context=Context.EVAL)
-def as_int(x: Any) -> Union[int, Iterable[int]]:
-    """Convert an object or elements of an iterable into int
-
-    Args:
-        x: The object
-
-    Returns:
-        When x is an array or a series, return x.astype(int).
-        When x is iterable, convert elements of it into ints
-        Otherwise, convert x to int.
-    """
-    if is_categorical_dtype(x):
-        return categorized(x).codes
-    return _as_type(x, int)
-
-@register_func(None, context=Context.EVAL)
-def as_integer(x: Any) -> Union[numpy.int64, Iterable[numpy.int64]]:
-    """Convert an object or elements of an iterable into int64
-
-    Args:
-        x: The object
-
-    Returns:
-        When x is an array or a series, return x.astype(numpy.int64).
-        When x is iterable, convert elements of it into numpy.int64s
-        Otherwise, convert x to numpy.int64.
-    """
-    if is_categorical_dtype(x):
-        return categorized(x).codes
-    return _as_type(x, numpy.int64)
-
-as_int64 = as_integer
-
-@register_func(None, context=Context.EVAL)
-def as_numeric(x: Any) -> NumericOrIter:
-    """Make elements numeric"""
-    try:
-        return as_integer(x)
-    except (ValueError, TypeError):
-        return as_double(x)
-
-@register_func(None, context=Context.EVAL)
-def as_logical(x: Any) -> BoolOrIter:
-    """Convert an object or elements of an iterable into bool
-
-    Args:
-        x: The object
-
-    Returns:
-        When x is an array or a series, return x.astype(bool).
-        When x is iterable, convert elements of it into bools
-        Otherwise, convert x to bool.
-    """
-    return _as_type(x, bool)
-
-as_bool = as_logical
-
-@register_func(None, context=Context.EVAL)
-def is_numeric(x: Any) -> BoolOrIter:
-    """Check if x is numeric type"""
-    if is_scalar(x):
-        return isinstance(x, (int, float, complex, numpy.number))
-    if isinstance(x, Series):
-        return is_numeric_dtype(x)
-    return Array([is_numeric(elem) for elem in x])
-
-@register_func(None, context=Context.EVAL)
-def is_character(x: Any) -> BoolOrIter:
-    """Mimic the is.character function in R
-
-    Args:
-        x: The elements to check
-
-    Returns:
-        True if
-    """
-    if is_scalar(x):
-        return isinstance(x, str)
-    return is_string_dtype(x)
-
-@register_func(None, context=Context.EVAL)
-def is_categorical(x: Any) -> bool:
-    """Check if x is categorical data"""
-    return is_categorical_dtype(x)
-
-is_factor = is_categorical
-
-@register_func(None, context=Context.EVAL)
-def is_int64(x: Any) -> BoolOrIter:
-    """Check if x is double/float data"""
-    if is_scalar(x):
-        return isinstance(x, (int, numpy.int64))
-    return is_int64_dtype(x)
-
-@register_func(None, context=Context.EVAL)
-def is_integer(x: Any) -> BoolOrIter:
-    """Check if x is double/float data"""
-    if is_scalar(x):
-        return is_scalar_int(x)
-    return is_integer_dtype(x)
-
-is_int = is_integer
-
-@register_func(None, context=Context.EVAL)
-def is_double(x: Any) -> BoolOrIter:
-    """Check if x is double/float data"""
-    if is_scalar(x):
-        return isinstance(x, (float, numpy.float))
-    return is_float_dtype(x)
-
-is_float = is_double
-
-@register_func(None, context=Context.UNSET)
-def c(*elems: Any) -> Collection:
-    """Mimic R's concatenation. Named one is not supported yet
-    All elements passed in will be flattened.
-
-    Args:
-        _data: The data piped in
-        *elems: The elements
-
-    Returns:
-        A collection of elements
-    """
-    return Collection(*elems)
-
-@register_func(None, context=Context.EVAL)
-def seq_along(
-        along_with: Iterable[Any],
-        _base0: Optional[bool] = None
-) -> SeriesLikeType:
-    """Generate sequences along an iterable"""
-    _base0 = get_option('index.base.0', _base0)
-    return Array(range(len(along_with))) + int(not _base0)
-
-@register_func(None, context=Context.EVAL)
-def seq_len(
-        length_out: IntOrIter,
-        _base0: Optional[bool] = None
-) -> SeriesLikeType:
-    """Generate sequences with the length"""
-    _base0 = get_option('index.base.0', _base0)
-    if is_scalar(length_out):
-        return Array(range(int(length_out))) + int(not _base0)
-    if len(length_out) > 1:
-        logger.warning(
-            "In seq_len(%r) : first element used of 'length_out' argument",
-            length_out
-        )
-    length_out = int(list(length_out)[0])
-    return Array(range(length_out)) + int(not _base0)
-
-
-@register_func(None, context=Context.EVAL)
-def seq(
-        from_: IntType = None,
-        to: IntType = None,
-        by: IntType = None,
-        length_out: IntType = None,
-        along_with: IntType = None,
-        _base0: Optional[bool] = None,
-) -> SeriesLikeType:
-    """Generate a sequence
-
-    https://rdrr.io/r/base/seq.html
-
-    Note:
-        This API is consistent with r-base's seq. 1-based and inclusive.
-    """
-    _base0 = get_option('index.base.0', _base0)
-    if along_with is not None:
-        return seq_along(along_with, _base0)
-    if from_ is not None and not is_scalar(from_):
-        return seq_along(from_, _base0)
-    if length_out is not None and from_ is None and to is None:
-        return seq_len(length_out)
-
-    base = int(not _base0)
-
-    if from_ is None:
-        from_ = base
-    elif to is None:
-        from_, to = base, from_
-
-    if length_out is not None:
-        by = (float(to) - float(from_)) / float(length_out - 1)
-
-    elif by is None:
-        by = 1 if to > from_ else -1
-        length_out = to - from_ + base if to > from_ else from_ - to + base
-    else:
-        length_out = (to - from_ + .1 * by + float(base) * by) // by
-    return Array([from_ + n * by for n in range(int(length_out))])
-
-@register_func(None, context=Context.EVAL)
-def abs(x: Any) -> NumericOrIter:
-    """Get the absolute value"""
-    return numpy.abs(x)
-
-@register_func(None, context=Context.EVAL)
-def ceiling(x: NumericOrIter) -> IntOrIter:
-    """Get the ceiling integer of x
-
-    Args:
-        x: The number
-
-    Returns:
-        The ceiling integer of x
-    """
-    if is_scalar(x):
-        return math.ceil(x)
-    return list(map(math.ceil, x))
-
-@register_func(None, context=Context.EVAL)
-def floor(x: NumericOrIter) -> IntOrIter:
-    """Get the floor integer of x
-
-    Args:
-        x: The number
-
-    Returns:
-        The ceiling integer of x
-    """
-    if is_scalar(x):
-        return math.floor(x)
-    return list(map(math.floor, x))
-
-def cumx(x: Any, how: str) -> numpy.ndarray:
-    """Cummulative along the elements in x
-
-    numpy and pandas' cums only put NA at the position of original NA, and skip
-    it for later elements. However, all elements afterwards are NA's in R.
-    """
-    if is_scalar(x):
-        x = [x]
-
-    if len(x) == 0:
-        return Array([])
-
-    out = []
-    out_append = out.append
-    def append(elem):
-        if is_null(out[-1]) or is_null(elem):
-            out_append(NA)
-        elif how == 'sum':
-            out_append(out[-1] + elem)
-        elif how == 'prod':
-            out_append(out[-1] * elem)
-        elif how == 'min':
-            out_append(builtins.min(out[-1], elem))
-        elif how == 'max':
-            out_append(builtins.max(out[-1], elem))
-
-    for i, elem in enumerate(x):
-        if i == 0:
-            out_append(elem)
-        else:
-            append(elem)
-    return Array(out)
-
-@register_func(None, context=Context.EVAL)
-def cumsum(x: Any) -> numpy.ndarray:
-    """Cummulative sum along elements in x"""
-    return cumx(x, 'sum')
-
-@register_func(None, context=Context.EVAL)
-def cumprod(x: Any) -> numpy.ndarray:
-    """Cummulative product along elements in x"""
-    return cumx(x, 'prod')
-
-@register_func(None, context=Context.EVAL)
-def cummin(x: Any) -> numpy.ndarray:
-    """Cummulative min along elements in x"""
-    return cumx(x, 'min')
-
-@register_func(None, context=Context.EVAL)
-def cummax(x: Any) -> numpy.ndarray:
-    """Cummulative max along elements in x"""
-    return cumx(x, 'max')
 
 @register_func(None, context=Context.EVAL)
 def cut(
@@ -573,205 +62,6 @@ def cut(
         ordered=ordered_result
     )
 
-@register_func(None, context=Context.EVAL)
-def sample(
-        x: Union[int, Iterable[Any]],
-        size: Optional[int] = None,
-        replace: bool = False,
-        prob: Optional[Iterable[NumericType]] = None
-) -> Iterable[Any]:
-    """Takes a sample of the specified size from the elements of x using
-    either with or without replacement.
-
-    https://rdrr.io/r/base/sample.html
-
-    Args:
-        x: either a vector of one or more elements from which to choose,
-            or a positive integer.
-        n: a positive number, the number of items to choose from.
-        size: a non-negative integer giving the number of items to choose.
-        replace: should sampling be with replacement?
-        prob: a vector of probability weights for obtaining the elements of
-            the vector being sampled.
-
-    Returns:
-        A vector of length size with elements drawn from either x or from the
-        integers 1:x.
-    """
-    if isinstance(x, str):
-        x = list(x)
-    if size is None:
-        size = len(x) if is_iterable(x) else x
-    return numpy.random.choice(x, int(size), replace=replace, p=prob)
-
-@register_func(None, context=Context.EVAL)
-def round(
-        number: NumericOrIter,
-        ndigits: int = 0
-) -> NumericOrIter:
-    """Rounding a number"""
-    if is_series_like(number):
-        return number.round(ndigits)
-    return builtins.round(number, ndigits)
-
-@register_func(None, context=Context.EVAL)
-def sqrt(x: Any) -> bool:
-    """Get the square root of a number"""
-    return numpy.sqrt(x)
-
-@register_func(None, context=Context.EVAL)
-def droplevels(x: Categorical) -> Categorical:
-    """drop unused levels from a factor
-
-    Args:
-        x: The categorical data
-
-    Returns:
-        The categorical data with unused categories dropped.
-    """
-    return categorized(x).remove_unused_categories()
-
-@register_func(None, context=Context.EVAL)
-def levels(x: CategoricalLikeType) -> Optional[List[Any]]:
-    """Get levels from a factor
-
-    Args:
-        x: The categorical data
-
-    Returns:
-        levels of the categorical data
-    """
-    if not is_categorical_dtype(x):
-        return None
-    if isinstance(x, Series):
-        x = x.cat
-    return x.categories
-
-@register_func(None, context=Context.EVAL)
-def rep(
-        x: Any,
-        times: Union[int, Iterable[int]] = 1,
-        length: Optional[int] = None, # pylint: disable=redefined-outer-name
-        each: int = 1
-) -> numpy.ndarray:
-    """replicates the values in x
-
-    Args:
-        x: a vector or scaler
-        times: number of times to repeat each element if of length len(x),
-            or to repeat the whole vector if of length 1
-        length: non-negative integer. The desired length of the output vector
-        each: non-negative integer. Each element of x is repeated each times.
-
-    Returns:
-        A list of repeated elements in x.
-    """
-    if is_scalar(x):
-        x = [x]
-    if is_iterable(times):
-        if len(times) != len(x):
-            raise ValueError(
-                "Invalid times argument, expect length "
-                f"{len(times)}, got {len(x)}"
-            )
-        if each != 1:
-            raise ValueError(
-                "Unexpected each argument when times is an iterable."
-            )
-
-    if is_scalar_int(times):
-        x = Array([elem for elem in x for _ in range(each)] * int(times))
-    else:
-        x = Array([elem for n, elem in zip(times, x) for _ in range(n)])
-    if length is None:
-        return x
-    repeats = length // len(x) + 1
-    x = numpy.tile(x, repeats)
-    return x[:length]
-
-@register_func(None, context=Context.EVAL)
-def rev(x: Iterable[Any]) -> numpy.ndarray:
-    """Get reversed vector"""
-    dtype = getattr(x, 'dtype', None)
-    return Array(list(reversed(x)), dtype=dtype)
-
-
-@register_func(None, context=Context.EVAL)
-def unique(x: Iterable[Any]) -> numpy.ndarray:
-    """Get unique elements"""
-    # return numpy.unique(x)
-    return pandas.unique(x) # keeps order
-
-length = register_func(None, context=Context.EVAL, func=length_of)
-
-@register_func(None, context=Context.EVAL)
-def lengths(x: Any) -> List[int]:
-    """Lengths of elements in x"""
-    if is_scalar(x):
-        return [1]
-    return [length(elem) for elem in x]
-
-# ---------------------------------
-# Plain functions
-# ---------------------------------
-
-def factor(
-        x: Optional[Iterable[Any]] = None,
-        # pylint: disable=redefined-outer-name
-        levels: Optional[Iterable[Any]] = None,
-        exclude: Any = NA,
-        ordered: bool = False
-) -> Categorical:
-    """encode a vector as a factor (the terms ‘category’ and ‘enumerated type’
-    are also used for factors).
-
-    If argument ordered is TRUE, the factor levels are assumed to be ordered
-
-    Args:
-        x: a vector of data
-        levels: an optional vector of the unique values (as character strings)
-            that x might have taken.
-        exclude: a vector of values to be excluded when forming the set of
-            levels. This may be factor with the same level set as x or
-            should be a character
-        ordered: logical flag to determine if the levels should be regarded
-            as ordered (in the order given).
-    """
-    if x is None:
-        x = []
-
-    if is_categorical_dtype(x):
-        x = x.to_numpy()
-    ret = Categorical(
-        x,
-        categories=levels,
-        ordered=ordered
-    )
-    if is_scalar(exclude):
-        exclude = [exclude]
-
-    return ret.remove_categories(exclude)
-
-def context(data: DataFrame) -> Any:
-    """Evaluate verbs, functions in the
-    possibly modifying (a copy of) the original data.
-
-    It mimic the `with` function in R, but you have to write it in a python way,
-    which is using the `with` statement. And you have to use it with `as`, since
-    we need the value returned by `__enter__`.
-
-    Args:
-        data: The data
-        func: A function that is registered by
-            `pipda.register_verb` or `pipda.register_func`.
-        *args: Arguments for func
-        **kwargs: Keyword arguments for func
-
-    Returns:
-        The original or modified data
-    """
-    return WithDataEnv(data)
-
 @register_func(None, context=None)
 def identity(x: Any) -> Any:
     """Return whatever passed in
@@ -793,55 +83,26 @@ def expandgrid(*args: Iterable[Any], **kwargs: Iterable[Any]) -> DataFrame:
         columns=iters.keys()
     )
 
-@register_func(None)
-def Re(numbers: NumericOrIter) -> numpy.ndarray:
-    """Real part of complex numbers"""
-    if is_scalar(numbers):
-        return numbers.real
-    return numpy.real(numbers)
+# ---------------------------------
+# Plain functions
+# ---------------------------------
 
-@register_func(None)
-def Im(numbers: NumericOrIter) -> numpy.ndarray:
-    """Imaginary part of complex numbers"""
-    if is_scalar(numbers):
-        return numbers.imag
-    return numpy.imag(numbers)
+def data_context(data: DataFrame) -> Any:
+    """Evaluate verbs, functions in the
+    possibly modifying (a copy of) the original data.
 
-
-# Logic ----------------------------------------------------
-
-@register_func(None, context=Context.EVAL)
-def istrue(x: Any) -> bool:
-    """Check if a value is a scalar True, like `isTRUE()` in `R`.
-
-    If the value is non-scalar, will return False
-    This does not require exact True, but a value that passes
-    `if` test in python
+    It mimic the `with` function in R, but you have to write it in a python way,
+    which is using the `with` statement. And you have to use it with `as`, since
+    we need the value returned by `__enter__`.
 
     Args:
-        x: The value to test
+        data: The data
+        func: A function that is registered by
+            `pipda.register_verb` or `pipda.register_func`.
+        *args: Arguments for func
+        **kwargs: Keyword arguments for func
 
     Returns:
-        True if x is True otherwise False
+        The original or modified data
     """
-    if not is_scalar(x):
-        return False
-    return bool(x)
-
-@register_func(None, context=Context.EVAL)
-def isfalse(x: Any) -> bool:
-    """Check if a value is a scalar False, like `isFALSE()` in `R`.
-
-    If the value is non-scalar, will return False
-    This does not require exact False, but a value that fails
-    `if` test in python
-
-    Args:
-        x: The value to test
-
-    Returns:
-        False if x is False otherwise True
-    """
-    if not is_scalar(x):
-        return False
-    return not bool(x)
+    return WithDataEnv(data)
