+++ datar/base/which.py
@@ -1,28 +1,13 @@
 """Checking an iterable against itself or another one"""
-import builtins
-from typing import Any, Iterable, Optional
+from typing import Iterable, Optional
 
 import numpy
 from pipda import register_func
 
-from ..core.types import BoolOrIter, is_scalar
 from ..core.utils import get_option
 from ..core.contexts import Context
 
-@register_func(None)
-def is_element(elem: Any, elems: Iterable[Any]) -> BoolOrIter:
-    """R's `is.element()` or `%in%`.
-
-    Alias `is_in()`
-
-    We can't do `a %in% b` in python (`in` behaves differently), so
-    use this function instead
-    """
-    if is_scalar(elem):
-        return elem in elems
-    return numpy.isin(elem, elems)
-
-@register_func(None)
+@register_func(None, context=Context.EVAL)
 def which(x: Iterable[bool], _base0: Optional[bool] = None) -> Iterable[int]:
     """Convert a bool iterable to indexes
 
@@ -68,15 +53,3 @@ def which_max(x: Iterable, _base0: bool = True) -> int:
         The index of the element with the minimum value
     """
     return numpy.argmax(x) + int(not get_option('which.base.0', _base0))
-
-# pylint: disable=invalid-name
-is_in = is_element
-
-# pylint: disable=unnecessary-lambda, redefined-builtin
-all = register_func(None, context=Context.EVAL)(
-    # can't set attributes to builtins.all, so wrap it.
-    lambda *args, **kwargs: builtins.all(*args, **kwargs)
-)
-any = register_func(None, context=Context.EVAL)(
-    lambda *args, **kwargs: builtins.any(*args, **kwargs)
-)
